pyproject.toml
[project]
name = "shakespeare-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "langchain-community>=0.3.27",
    "langchain[openai]>=0.3.27",
]
